% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{add_name_species_in_importation}
\alias{add_name_species_in_importation}
\title{Add the name of the species to the list containing the names and lengths of
chromosomes.}
\usage{
add_name_species_in_importation(
  path,
  list_chromosome_lengths,
  vect_files_to_import
)
}
\arguments{
\item{path}{This string contains the path to the file to consider (one file
for one species).}

\item{list_chromosome_lengths}{List containing one level for each species.
Each level contains one data frame with the name and length of the chromosomes
of this species.}

\item{vect_files_to_import}{This vector contains all the paths to use to
import the chromosome names and lengths.}
}
\value{
A data frame containing the names and lengths of chromosomes for
one species with the name of the species inside it.
}
\description{
This function uses the list of files to import to know which level of the
list containing a data frame with the name and length of the chromosomes
to consider. Once this is done, it adds the name of the species to the
corresponding table of the input list.
}
\examples{
add_name_species_in_importation("Jaera_albifrons_chromosomes.fasta.fai",
                                list(data.frame(
                                     "Chromosome" = c("Chrom_1", "Chrom_2"),
                                     "Length" = c(10, 20)),
                                     data.frame(
                                     "Chromosome" = c("Chrom_1", "Chrom_2"),
                                     "Length" = c(12, 25))
                                     ),
                                c("Jaera_albifrons_chromosomes.fasta.fai",
                                "Jaera_praehirsuta_chromosomes.fasta.fai"))
}
