% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{add_not_attributed_chromosomes}
\alias{add_not_attributed_chromosomes}
\title{Adds chromosomes that were not added in the \code{find_correspondences_for_rearrangements}
function.}
\usage{
add_not_attributed_chromosomes(
  information_rearrangements,
  chroms_not_attributed
)
}
\arguments{
\item{information_rearrangements}{This data frame contains the top two
correspondences for each chromosome (two chromosomes that have the largest
cumulative matching sequence).}

\item{chroms_not_attributed}{This vector contains the names of the chromosomes
that did not get attributed to another chromosome yet.}
}
\value{
A data frame containing the correspondence between the chromosomes
that were not attributed and their most corresponding chromosome in another
species.
}
\description{
This function selects the column of the input containing chromosomes that
were not added to the list of correspondences between chromosomes and adds
them to the list of correspondences with the chromosome that corresponds
the most.
}
