% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{add_table_to_df_in_iteration}
\alias{add_table_to_df_in_iteration}
\title{Adds a data frame to another in an iteration}
\usage{
add_table_to_df_in_iteration(df1, df2)
}
\arguments{
\item{df1}{(data.frame).
This data frame is the name of the data frame the data will be added to}

\item{df2}{(data.frame).
This data frame contains the data to add to the df1.}
}
\value{
(data.frame).
This is the binding of the two data frames
}
\description{
In an iteration, if we want to rbind a data frame to another, we have to
check if the data frame is empty and if it is not, we can rbind the two
data frames together
}
\examples{
data1 <- data.frame("a" = c(1, 2, 3), "b" = c(5, 6, 7))
data2 <- data.frame("a" = c(4, 5, 6), "b" = c(1, 2, 3))
full_data <- add_table_to_df_in_iteration(data1, data2)
}
