% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alignments_ready_to_plot}
\alias{alignments_ready_to_plot}
\title{Alignments ready to plot}
\format{
\subsection{\code{alignments_ready_to_plot}}{

A data frame with 27,749 rows and 9 columns:
\describe{
\item{qname}{Name of the assembled chromosome for the query species.}
\item{qstart, qend}{Start/End coordinates of the mapping sequence on the
chromosome of the query sepcies.}
\item{tname}{Name of the assembled chromosome for the target species.}
\item{tstart, tend}{Start/End coordinates of the mapping sequence on the
chromosome of the target species.}
\item{alen}{Length of the mapping sequence.}
\item{query, target}{Names of the query/target species.}
}
}
}
\source{
Run of \code{minimap2} on chromosome-scale reference genomes published by
the Darwin Tree of Life project on which we reordered chromosomes that were
not assembled in the same direction.
}
\usage{
alignments_ready_to_plot
}
\description{
An example of a dataset that is ready to plot without having to prepare it
in the examples. We simply return chromosomes that were assembled differently
for the studied species.
}
\keyword{datasets}
