% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_alignments}
\alias{all_alignments}
\title{All alignments}
\format{
\subsection{\code{all_alignments}}{

A data frame with 27,749 rows and 9 columns:
\describe{
\item{qname}{Name of the assembled chromosome for the query species.}
\item{qstart, qend}{Start/End coordinates of the mapping sequence on the
chromosome of the query species.}
\item{tname}{Name of the assembled chromosome for the target species.}
\item{tstart, tend}{Start/End coordinates of the mapping sequence on the
chromosome of the target species.}
\item{alen}{Length of the mapping sequence.}
\item{query, target}{Names of the query/target species.}
}
}
}
\source{
Run of \code{minimap2} on chromosome-scale reference genomes published by
the Darwin Tree of Life project.
}
\usage{
all_alignments
}
\description{
An example of a paf file imported as a data frame that can be used in tidyverse.
}
\keyword{datasets}
