% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{any_string_in_vector_contains_pattern}
\alias{any_string_in_vector_contains_pattern}
\title{Checks if any string in a vector contains a pattern}
\usage{
any_string_in_vector_contains_pattern(pattern, vec_to_check)
}
\arguments{
\item{pattern}{Pattern to look for in the vector}

\item{vec_to_check}{Vector in which to look for the pattern}
}
\value{
Boolean: \code{TRUE} if the pattern is in any of the vector's string and
\code{FALSE} otherwise.
}
\description{
Checks if any string in the vector contains the specified pattern
}
\examples{
pattern_to_look_for <- "th"
vect1 <- c("cameleon", "panther", "tea", "thermos", "text")
any_string_in_vector_contains_pattern(pattern_to_look_for, vect1)
}
