% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{are_all_in}
\alias{are_all_in}
\title{Check if all occurrences of the first vector are in the second one}
\usage{
are_all_in(vec1, vec2)
}
\arguments{
\item{vec1}{The vector to test}

\item{vec2}{The vector in which to test}
}
\value{
boolean TRUE if all the occurrences of the first vector are also in
the second one and FALSE otherwise.
}
\description{
Checks if all the occurrences of the first
vector are also contained in the second one
}
\examples{
vect1 <- c("a", "b", "c", "d")
vect2 <- c("z", "x", "y", "a")
vect3 <- c("z", "a", "x", "b", "y", "c", "d")
are_all_in(vect1, vect3)
are_all_in(vect2, vect3)
are_all_in(vect1, vect2)
}
