% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{arrange_for_polygons}
\alias{arrange_for_polygons}
\title{Arrange mapping positions along the genome to be able to plot them in the
synteny plot.}
\usage{
arrange_for_polygons(positions_to_plot)
}
\arguments{
\item{positions_to_plot}{This data frame contains all the information ready
to be plotted along the genome.}
}
\value{
The same data frame, but with a specific order for each species.
}
\description{
This function arranges the positions of the mapping coordinates in opposite
orders for successive lines of the final plot for it to plot polygons from
the start to the end of the mapping fragments.
}
\examples{
# Import data
data(data_to_plot)
data_to_plot

# Test the function
arrange_for_polygons(data_to_plot)
}
