% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromosome_names}
\alias{chromosome_names}
\title{Chromosome names}
\format{
\subsection{\code{chromosome_names}}{

A data frame with 15 rows and 5 columns
\describe{
\item{Chromosome}{Assembly name of the chromosome.}
\item{Chromosome_name}{Alignment name of the chromosome.}
\item{Species}{Species name.}
\item{Length}{Length of the chromosome in base pairs.}
\item{Cumul_position_start_chromosome}{Cumulative position at which the
chromosome starts if they are arranged from chromosome 1 to chromosome n.
This cumulative position is independent for the different species and is
counted in base pairs.}
}
}
}
\usage{
chromosome_names
}
\description{
Assembly and alignment names of the chromosomes as well as their lengths and
cumulative positions for each species.
}
\keyword{datasets}
