% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{compute_cumulative_positions_along_genome}
\alias{compute_cumulative_positions_along_genome}
\title{Compute the cumulative positions of the correspondences along the genome}
\usage{
compute_cumulative_positions_along_genome(
  temp_data_to_plot,
  name_of_chromosomes
)
}
\arguments{
\item{temp_data_to_plot}{A data frame containing all the correspondences and
might have been filtered a little bit.}

\item{name_of_chromosomes}{This data frame contains the names, lengths, and
cumulative positions of the start of chromosomes along the genome for each
considered species.}
}
\value{
A data frame containing the correspondences that are to be plotted
with the cumulative positions of the correspondences along the genome of the
different species.
}
\description{
This function computes the cumulative positions of the correspondences
along the genomes of the different species.
}
\examples{
# Prepare data
example_data_to_plot <- data.frame(
 "qname" = c("LG1", "LG1", "LG2", "LG2"),
 "tname" = c("LG1", "LG1", "LG2", "LG2"),
 "qstart" = c(2, 40, 20, 60),
 "qend" = c(25, 77, 50, 80),
 "tstart" = c(5, 42, 15, 66),
 "tend" = c(22, 72, 45, 88),
 "query" = rep("praehirsuta", 4),
 "target" = rep("albifrons", 4)
)

example_names_of_chromosomes <- data.frame(
 "Chromosome" = c("LG1", "LG2", "LG1", "LG2"),
 "Length" = c(100, 150, 98, 120),
 "Cumul_position_start_chromosome" = c(0, 101, 0, 99),
 "Species" = c(rep("albifrons", 2), rep("praehirsuta", 2))
)

# Test the function
compute_cumulative_positions_along_genome(example_data_to_plot,
                                          example_names_of_chromosomes)
}
