% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{contains_dot}
\alias{contains_dot}
\title{Checks if there is a dot in the string}
\usage{
contains_dot(string)
}
\arguments{
\item{string}{String in which to look for a dot}
}
\value{
Boolean: \code{TRUE} if the pattern contains a dot and \code{FALSE} otherwise
}
\description{
This function uses the \code{grepl} function to check the string for a dot
}
\examples{
text1 <- "This is fun."
text2 <- "how you doin"
contains_dot(text1)
contains_dot(text2)
}
