% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{contains_hyphen}
\alias{contains_hyphen}
\title{Checks if there is a hyphen in the string}
\usage{
contains_hyphen(string)
}
\arguments{
\item{string}{String in which to look for a hyphen}
}
\value{
Boolean: \code{TRUE} if the pattern contains a hyphen and \code{FALSE} otherwise
}
\description{
This function uses the \code{grepl} function to check the string for a hyphen
}
\examples{
text1 <- "This is fun."
text2 <- "main-text"
contains_hyphen(text1)
contains_hyphen(text2)
}
