% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{convert_to_numeric_if_possible}
\alias{convert_to_numeric_if_possible}
\title{Try to convert a column to numeric if it is possible}
\usage{
convert_to_numeric_if_possible(x)
}
\arguments{
\item{x}{This is a column from a dataframe}
}
\value{
The column as numeric or as it was.
}
\description{
This function tries to convert the values of the specified column
to numeric. If this returns NA (expected for characters or booleans), it
keeps the column as it is.
}
\examples{
df1 <- data.frame("numbers_in_text" = c("1", "2", "3"), "text" = c("you", "can", "test"))
df1
df1$numbers_in_text <- convert_to_numeric_if_possible(df1$numbers_in_text)
df1
}
