% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{correspondences_chromosomes}
\alias{correspondences_chromosomes}
\title{Correspondences between chromosomes}
\format{
\subsection{\code{correspondences_chromosomes}}{

A data frame with 13 rows and 4 columns.
\describe{
\item{query, target}{Names of the query/target species.}
\item{qname, tname}{Names of the assembled chromosomes of the query/target
species.}
}
}
}
\usage{
correspondences_chromosomes
}
\description{
Correspondences between chromosomes of the different studied species.
}
\keyword{datasets}
