% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_to_plot}
\alias{data_to_plot}
\title{Data to plot}
\format{
\subsection{\code{data_to_plot}}{

A data frame with 4 rows and 7 columns.
\describe{
\item{Chromosome_name_query, Chromosome_name_target}{Alignment names of the
chromosomes of the query/target species.}
\item{Position}{Group number to trace the polygon in the "make_synteny_plot"
function.}
\item{name}{Column resulting from the use of a pivot_longer function that
is used to be able to have all the coordinates of the polygon to trace in
the synteny plot in one unique column.}
\item{delims_polygon}{Coordinates of the mapping sequence along the
chromosomes of the query/target species.}
\item{direction}{Indication for the "arrange_for_polygons" function to order
the data in the correct order for polygons to be traced.}
\item{Species}{Target/query species.}
}
}
}
\usage{
data_to_plot
}
\description{
Example data to plot in the "make_synteny_plot" function of the package.
}
\keyword{datasets}
