% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_rename_chromosomes}
\alias{example_rename_chromosomes}
\title{Example of chromosomes to rename}
\format{
\subsection{\code{example_rename_chromosomes}}{

A data frame with 25 rows and 3 columns.
\describe{
\item{Chromosome}{Assembly name of the chromosome.}
\item{Chromosome_name}{Alignment name of the chromosome.}
\item{Species}{Species name.}
}
}
}
\usage{
example_rename_chromosomes
}
\description{
Example of data frame in which we have some chromosomes to rename in the
package.
}
\keyword{datasets}
