% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{find_correspondences_for_rearrangements}
\alias{find_correspondences_for_rearrangements}
\title{Get the names of the corresponding chromosomes when there are some complex
rearrangements and the name of the chromosome that matches most is not enough}
\usage{
find_correspondences_for_rearrangements(information_rearrangements)
}
\arguments{
\item{information_rearrangements}{This data frame contains the top two
chromosomes of species B that matches each chromosome of species A.}
}
\value{
A data frame that contains all the names of chromosomes from
species A and the name of the corresponding chromosome of species B.
}
\description{
This function takes the top two chromosomes of species B that match the
chromosome in species A and counts the number of times these top two
chromosomes of species B are the same. If the pair of chromosomes of species
B is unique, the first chromosome of species B (the one with the largest
cumulative mapping sequence) is given as correspondence to the chromosome
of species A. If the pair of chromosomes of species B is present more than
once, it checks the order of magnitude of the cumulative mapping sequence
for the top two chromosomes of species B. If the order of magnitude is
not the same, the first chromosome of species B (the one with the largest
cumulative mapping sequence) is given as correspondence. If the order of
magnitude is the same, the function attributes the first chromosome of
species A to the first chromosome of species B (the one with the largest
cumulative mapping sequence) and attributes the remaining chromosome of
species A to the remaining chromosome of species B.
}
\examples{
# Import data
data(information_rearrangements)

find_correspondences_for_rearrangements(information_rearrangements)

}
