% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_chromosome_names_all_species}
\alias{get_chromosome_names_all_species}
\title{Get the names of all the chromosomes for all the species that are specified}
\usage{
get_chromosome_names_all_species(
  all_chromosomes,
  correspondences_chromosomes,
  all_alignments,
  reference_chromosomes = NULL
)
}
\arguments{
\item{all_chromosomes}{This data frame contains at least the name of the
chromosomes that were given when the chromosomes were assembled, and to which
species this assembly corresponds.}

\item{correspondences_chromosomes}{This data frame contains four columns:
the species of the query and target species and the names of the corresponding
chromosomes from these two species. This table is the output of the \code{get_corresponding_chromosomes_species}
function.}

\item{all_alignments}{This data frame contains all the coordinates of the aligned
genomes on one another.}

\item{reference_chromosomes}{This data fame contains the name of the species
used as reference, the names of the chromosomes when assembled and the names
of the chromosomes in the data analysis. The default value of this argument
is \code{NULL}. If this is the case, a default species will be chosen as reference.}
}
\value{
A table with at least three columns: the names
of all species, the names of all chromosomes that were given when assembled
and the names of the chromosomes given in the data analysis. Any supplementary
column contained in the \code{all_chromosomes} table will also be found in the
output table.
}
\description{
This function uses a species as reference to name the chromosomes of all the species
in the same way. For example, if we have two species, A and B, and we
know that chromosome 1 of species A corresponds to chromosome OZ1292456
from species B, this function will allow to name the chromosome OZ1292456
as Chromosome 1 in species B. This function automates this process for all
the pairs of species that are aligned on each other.
}
\examples{
# Import paf files
data(all_alignments)

# Import the lengths and assembly names of chromosomes
data(names_of_chromosomes)

# Get the correspondences between chromosomes of different species
correspondences_chromosomes <- get_corresponding_chromosomes_species(
   all_alignments, names_of_chromosomes
)
data(reference_chromosomes)


get_chromosome_names_all_species(names_of_chromosomes,
                                 correspondences_chromosomes,
                                 all_alignments,
                                 reference_chromosomes)
}
