% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_chromosome_names_using_reference}
\alias{get_chromosome_names_using_reference}
\title{Get the name of the chromosomes using the names of one species as a reference}
\usage{
get_chromosome_names_using_reference(
  correspondences_chromosomes,
  reference_chromosomes,
  pair_of_species,
  reference_species
)
}
\arguments{
\item{correspondences_chromosomes}{This data frame contains the
correspondences between the chromosomes of the focal species and the
reference species.}

\item{reference_chromosomes}{This data frame contains three columns: the
name of the reference species, the names of the chromosomes given when they
were assembled and the name of the chromosomes that are given when doing the
analysis.}

\item{pair_of_species}{This string contains the name of the focal species
and the name of the reference species separated by a "_".}

\item{reference_species}{String which is the name of the reference species}
}
\value{
A table with three columns: the name of the focal species, the name
of the chromosomes of this species when they were assembled and the new name
given when doing the data analysis.
}
\description{
This function allows to name the chromosomes of one species
using the names given to the chromosomes of a reference species. We use
correspondences between the chromosome of the reference species and the
chromosome of the focal species to return the correct names.
}
\examples{
data(correspondences_chromosomes)
data(reference_chromosomes)

pair_of_species <- "ischiosetosa_albifrons"
reference_species <- unique(reference_chromosomes$Species)

get_chromosome_names_using_reference(correspondences_chromosomes,
                                     reference_chromosomes,
                                     pair_of_species,
                                     reference_species)
}
