% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_corresponding_chromosomes_species}
\alias{get_corresponding_chromosomes_species}
\title{Get the correspondences between chromosomes of different species}
\usage{
get_corresponding_chromosomes_species(all_alignments, all_chromosomes)
}
\arguments{
\item{all_alignments}{Get the correspondences between chromosomes of different species}

\item{all_chromosomes}{A data frame containing the names and lengths
of the different chromsomes for the different species.}
}
\value{
A table with four columns: two columns with the names of the species
the name of the chromosome in the first species and the name of the
corresponding chromosome in the second species
}
\description{
This function uses the alignments of the reference genomes got
from the "paf" files to know which chromosomes are the closest, thus
which chromosome will be called "Chrom1" in the different species.
To do this, we get the total length of aligned sequences per pair of
chromosomes and filter it to keep only the pair of chromosomes for which
the total length of aligned sequence is the greatest.
}
\examples{
 # Import data
data(all_alignments)
all_alignments

data(names_of_chromosomes)

# Test the function
get_corresponding_chromosomes_species(all_alignments, names_of_chromosomes)
}
