% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_names_of_fused_chromosomes}
\alias{get_names_of_fused_chromosomes}
\title{Separates and distributes a prefix separated by a dot to two suffixes separated
by hyphens}
\usage{
get_names_of_fused_chromosomes(string)
}
\arguments{
\item{string}{The string to be split}
}
\value{
A list with two levels: \code{name_first_chrom} is the prefix (part
before the dot) of the string, and \code{name_second_chrom} is the prefix with
the second part of the string (after the hyphen).
}
\description{
This function splits the string using a hyphen as delimiter, splits the string
using a dot as delimiter and returns the first part of the string (before
the hyphen) and the second part of the string (after the hyphen) with the
suffix (part after the dot).
}
\examples{
get_names_of_fused_chromosomes("LG1.a-b")
}
