% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_names_unfused_chromosomes}
\alias{get_names_unfused_chromosomes}
\title{Separates and distributes a prefix separated by a dot to a suffix and the
following letter of the alphabet.}
\usage{
get_names_unfused_chromosomes(string)
}
\arguments{
\item{string}{The string to be split}
}
\value{
A list with two levels: \code{name_first_chrom} is the string given as
input, and \code{name_second_chrom} is the prefix (part of the string before the
dot) with the next letter of the alphabet.
}
\description{
This function splits the string using a dot as delimiter, identifies which
letter constitutes the suffix and returns the string as well as the prefix
pasted with the next letter of the alphabet.
}
\examples{
get_names_unfused_chromosomes("LG1.a")
}
