% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_new_reference_species}
\alias{get_new_reference_species}
\title{Get a new reference species in the ones that were already run}
\usage{
get_new_reference_species(sps, already_done_species)
}
\arguments{
\item{sps}{Vector containing two species names to test}

\item{already_done_species}{vector containing all the species for which the
chromosome naming was already done}
}
\value{
the name of the new reference species to use
}
\description{
This function checks if the first species given has already been run. If
it is the case, it uses this as a new reference. Otherwise, it uses the
second species specified.
}
\examples{
get_new_reference_species(c("albifrons", "ischiosetosa"), c("albifrons"))
get_new_reference_species(c("albifrons", "ischiosetosa"), c("praehirsuta"))
}
