% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_species_number_according_to_order}
\alias{get_species_number_according_to_order}
\title{Get the number of the species according to a specified order}
\usage{
get_species_number_according_to_order(temp_data_to_plot, order_species)
}
\arguments{
\item{temp_data_to_plot}{This data frame contains all the sequences that map
on the chromosomes of different species that will be plotted.}

\item{order_species}{This named vector contains the names of the species and
the according number. The numbers of the species are the names of the vector
and the names of the species are the contents. The vector should be as follows:
\code{c("1" = "name_species_A", "2" = "name_species_B")}.}
}
\value{
A data frame containing the newly added species number.
}
\description{
To plot the species, as we want to add polygons to delimit the chromosomes,
we need to use a continuous scale. To do this, we use unique correspondences
between the species name and a unique number given to the species. This
function uses the specified vector to give the right number to the right
species.
}
\examples{
example_data_to_plot <- data.frame(
  "Species" = c("praehirsuta", "praehirsuta", "albifrons", "albifrons"),
  "data_to_plot" = c(1, 1, 4, 4)
)

species_order_in_plot <- c("1" = "praehirsuta", "2" = "albifrons")

get_species_number_according_to_order(example_data_to_plot,
                                      species_order_in_plot)
}
