% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Not in}
\usage{
x \%not_in\% y
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
A boolean: \code{TRUE} for each value of x that is NOT in y and \code{FALSE}
for every value of x that is in y.
}
\description{
This function is the opposite of the "\%in\%" function. It checks if the
values contained in input x are NOT contained in y.
}
\examples{
a <- c(1, 3, 5, 9)
b <- c(1, 4, 6, 8)
a \%not_in\% b
}
