% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{import_all_chromosome_lengths}
\alias{import_all_chromosome_lengths}
\title{Import the chromosome lengths for all files from one folder}
\usage{
import_all_chromosome_lengths(paths_to_genomes)
}
\arguments{
\item{paths_to_genomes}{Data frame containing the paths to the reference genomes
and the names of the species to which the reference genomes correspond.}
}
\value{
A data frame containing the names of the species, the names of the
chromosomes when assembled and the lengths of the chromosomes.
}
\description{
This function imports all the lengths of chromosomes from all the "fai" or
files in a directory
}
\examples{
data(paths_to_genomes)
paths_to_genomes

import_all_chromosome_lengths(paths_to_genomes)
}
