% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{import_chromosome_length_from_dict}
\alias{import_chromosome_length_from_dict}
\title{Imports the lengths of the chromosomes using the ".dict" file from the
reference genome indexing.}
\usage{
import_chromosome_length_from_dict(path)
}
\arguments{
\item{path}{This string of characters is the path to the ".dict" file}
}
\value{
A table containing the name and length of each chromosomes.
}
\description{
This function takes as input the path to the ".dict" file
obtained when the reference genome is indexed and keeps only the name
of the chromosome and its length.
}
\examples{
dict_file <- "TidySynt/extdata/refgenome/Jaera_albifrons_chromosomes.dict"
import_chromosome_length_from_dict(dict_file)
}
