% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{import_chromosome_length_from_fai}
\alias{import_chromosome_length_from_fai}
\title{Imports the lengths of the chromosomes using the ".fai" file from the
reference genome indexing using \code{samtools}.}
\usage{
import_chromosome_length_from_fai(path)
}
\arguments{
\item{path}{This string of characters is the path to the ".fai" file}
}
\value{
A table containing the name and length of each chromosomes.
}
\description{
This function takes as input the path to the ".fai" file
obtained when the reference genome is indexed and keeps only the name
of the chromosome and its length.
}
\examples{
fai_file <- "TidySynt/extdata/refgenome/Jaera_albifrons_chromosomes.fasta.fai"
import_chromosome_length_from_fai(fai_file)
}
