% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{import_chromosome_lengths_for_all_files_of_folder}
\alias{import_chromosome_lengths_for_all_files_of_folder}
\title{Import the chromosome lengths for all files from one folder}
\usage{
import_chromosome_lengths_for_all_files_of_folder(
  dir_name,
  preferences = "fai"
)
}
\arguments{
\item{dir_name}{Path to the directory that contains the "dict" or "fai" files}

\item{preferences}{default = \code{"fai"}. This string allows the function to know
if it should import preferentially the lengths from the \code{fai} or the \code{dict}
files.}
}
\value{
A data frame containing the names of the species, the names of the
chromosomes when assembled and the lengths of the chromosomes.
}
\description{
This function imports all the lengths of chromosomes from all the "fai" or
"dict" files in a directory
}
\examples{
import_chromosome_lengths_for_all_files_of_folder("TidySynt/extdata/refgenome/",
                                                  preferences = "fai")
import_chromosome_lengths_for_all_files_of_folder("TidySynt/extdata/refgenome/",
                                                  preferences = "dict")
}
