% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{import_paf}
\alias{import_paf}
\title{Import a "paf" file and transform it into a tidy table}
\usage{
import_paf(path)
}
\arguments{
\item{path}{This string of characters is the path to the \code{paf} file.}
}
\value{
A table with the contents of the \code{paf} file.
}
\description{
This function uses the \code{read_paf} function from the \code{pafr}
package to import a \code{.paf} file and transforms the list into a tidy table
that can me modified or plotted with the tidy format.
}
\examples{
import_paf("TidySynt/extdata/paf/albifrons_aligned_on_praehirsuta.paf")
}
