% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{information_rearrangements}
\alias{information_rearrangements}
\title{Top two correspondences for each chromosome of the query species and the total
mapping sequence length of these two top correspondences.}
\format{
\subsection{\code{information_rearrangements}}{

A data frame with 12 rows and 8 columns.
\describe{
\item{query, target}{Name of the query/target species.}
\item{qname}{Assembly name of the chromosomes for the query species.}
\item{sp_pairs}{Pair of query-target species separated by an underscore.}
\item{tname_1, tname_2}{Assembly names of the chromosomes that have the
largest/second largest cumulative mapping sequence with the chromosome of
the query species.}
\item{total_length_1, total_length_2}{Total cumulative mapping sequence
of the chromosome_1/_2 on the chromosome of the query species.}
}
}
}
\usage{
information_rearrangements
}
\description{
Top two correspondences for each chromosome of the query species and the total
mapping sequence length of these two top correspondences.
}
\keyword{datasets}
