% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{is_chromosome_pair_reversed}
\alias{is_chromosome_pair_reversed}
\title{Checks if the pair of chromosome is reversed or not.}
\usage{
is_chromosome_pair_reversed(all_alignments, chromosome_1, chromosome_2)
}
\arguments{
\item{all_alignments}{This data frame contains all the coordinates of the aligned
genomes on one another.}

\item{chromosome_1}{This string is the name of the first chromosome}

\item{chromosome_2}{This string is the name of the second chromosome}
}
\value{
Boolean. TRUE if the slope of the linear regression is negative, i.e.
if the chromosomes are reversed and FALSE if the chromosomes are well arranged.
}
\description{
This function filters the data to keep the correspondences
between the two specified chromosomes and runs a linear regression on the
corresponding positions in the two chromosomes to know if they are reversed
}
\examples{
data(all_alignments)

is_chromosome_pair_reversed(all_alignments, "OZ349671", "OZ242396")
}
