% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{is_same_order_of_magnitude}
\alias{is_same_order_of_magnitude}
\title{Checks if two numbers have the same order of magnitude}
\usage{
is_same_order_of_magnitude(a, b)
}
\arguments{
\item{a}{This is a number.}

\item{b}{This is a number.}
}
\value{
Boolean: \code{TRUE} if the two numbers have the same order of
magnitude and \code{FALSE} otherwise.
}
\description{
This function computes the order of magnitude of the two provided numbers
using the \code{order_magnitude} function provided above.
}
\examples{
is_same_order_of_magnitude(10, 50)
is_same_order_of_magnitude(50, 2)
}
