% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{iterate_over_pairs_of_species_to_get_name_of_chromosomes}
\alias{iterate_over_pairs_of_species_to_get_name_of_chromosomes}
\title{Go over all the pair of species given as parameter to name the chromosomes
according to a reference species}
\usage{
iterate_over_pairs_of_species_to_get_name_of_chromosomes(
  pair_of_species_to_do,
  already_done_pairs_of_species,
  named_chromosomes_all_sps,
  correspondences_chromosomes,
  reference_chromosomes,
  reference_species = NULL
)
}
\arguments{
\item{pair_of_species_to_do}{Vector containing all the pairs of species to
iterate over in the function. The pairs of species must be under the
following format: "sp1_sp2"}

\item{already_done_pairs_of_species}{Vector containing all the species for
which the chromosomes are already named. In this vector, each component is
the name of one species.}

\item{named_chromosomes_all_sps}{Data frame containing the name that was
given to a chromosome when it was assembled, the name of the chromosome that
is given during the analysis and the name of the species in which this
naming was done. If it is the first time that the function is run, this can
be an empty data frame.}

\item{correspondences_chromosomes}{Data frame containing four columns:
the names of the query and target species and the names of their
corresponding chromosomes. This is the output of the function \code{get_corresponding_chromosomes_species}.}

\item{reference_chromosomes}{Data frame containing three columns: the name
of the species used as reference, the name of the chromosomes for this reference
species when it was assembled and the name of the chromosomes that were
given in the data analysis.}

\item{reference_species}{Name of the reference species.}
}
\value{
list of three parameters:
1- named_chromosomes_all_sps: Data frame containing the name that was
given to a chromosome when it was assembled, the name of the chromosome that
is given during the analysis and the name of the species in which this
naming was done
2-already_done: vector containing the list of the species that were already done
3-to_do_later: vector containing the pairs of species to run in another call to the function
}
\description{
This function uses a for loop to iterate over all the combination of species
that are given in input to name all the corresponding chromosomes between
pairs of species. A reference can be specified, but is not necessarily
needed.
This function is called recursively, which is why some of its outputs are
also given in input. If you run this function for the first time, the
arguments that are given in the output can be given as empty.
}
