% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{make_names_unfused_chromosomes}
\alias{make_names_unfused_chromosomes}
\title{Make names for unfused chromosomes}
\usage{
make_names_unfused_chromosomes(name_chrom_given)
}
\arguments{
\item{name_chrom_given}{string that contains the name of the chromosome (for
example "LG1").}
}
\value{
A list with two levels: \code{name_first_chrom} is the name given as
input pasted ".a", and \code{name_second_chrom} is the name given as input pasted
with ".b".
}
\description{
Takes the name of the chromosome that is specified in input and returns the
name pasted with ".a" and ".b".
}
\examples{
make_names_unfused_chromosomes("LG1")
}
