% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_of_chromosomes}
\alias{names_of_chromosomes}
\title{Assembly names of the chromosomes.}
\format{
\subsection{\code{names_of_chromosomes}}{

A data frame with 24 rows and 3 columns.
\describe{
\item{Chromosome}{Assembly name of the chromosome.}
\item{Length}{Length of the chromosome.}
\item{Species}{Species name.}
}
}
}
\usage{
names_of_chromosomes
}
\description{
Assembly names of the chromosomes, their length and species.
}
\keyword{datasets}
