% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{order_magnitude}
\alias{order_magnitude}
\title{Get the order of magnitude of a number}
\usage{
order_magnitude(x)
}
\arguments{
\item{x}{This is a number.}
}
\value{
An integer.
}
\description{
This function gets the order of magnitude of a number using the \code{floor}
and \code{log10} functions.
}
\examples{
order_magnitude(1000.2)
order_magnitude(-500)
}
