% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paths_to_genomes}
\alias{paths_to_genomes}
\title{Paths to access reference genomes of different species}
\format{
\subsection{\code{paths_to_genomes}}{

A data frame with 2 rows and 2 columns
\describe{
\item{Species}{Species name.}
\item{paths}{Paths to the reference genome of each species}
}
}
}
\usage{
paths_to_genomes
}
\description{
Example of a data frame that stores the paths to the reference genomes of
different species that are specified here.
}
\keyword{datasets}
