% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{prepare_data_for_plotting}
\alias{prepare_data_for_plotting}
\title{Prepare the alignments to be plotted.}
\usage{
prepare_data_for_plotting(
  alignments,
  name_of_chromosomes,
  only_chromosome,
  small_alignment_threshold
)
}
\arguments{
\item{alignments}{This table contains the alignments for the different species
combinations. It is the equivalent of the \code{.paf} files, but transformed as tables.
It is the output of the \code{import_paf_from_folder} function.}

\item{name_of_chromosomes}{This table contains the names of the chromosomes
for the different species. It is the output of the \code{get_chromosome_names_all_species}
function}

\item{only_chromosome}{This argument is a string or vector of strings that
contains the name of one chromosome if you want to look at it individually.
It allows to filter the data to look only at this chromosome.}

\item{small_alignment_threshold}{This number allows to filter small fragments
that map. It defines the threshold of what the user specifies as "small".}
}
\value{
A list with two levels, first the filtered alignments and second the
filtered names of the chromosomes.
}
\description{
This function filters the alignments on the specified threshold and keeps
the specified chromosomes if any are specified. Otherwise, it keeps them
all.
}
\examples{
# Preapare and load data
data(alignments_ready_to_plot)
data(chromosome_names)

# Test the function
prepare_data_for_plotting(alignments_ready_to_plot, chromosome_names,
                          only_chromosome = NULL, 1e4)
# or if you wnat to select only one chromosome
prepare_data_for_plotting(alignments_ready_to_plot, chromosome_names,
                          only_chromosome = "LG_1$", 1e4)
}
