% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_chromosomes}
\alias{reference_chromosomes}
\title{Names of reference chromosomes}
\format{
\subsection{\code{reference_chromosomes}}{

A data frame with 12 rows and 3 columns.
\describe{
\item{Chromosome}{Assembly name of the chromosome.}
\item{Chromosome_name}{Alignment name of the chromosome for the reference
species.}
\item{Species}{Species name.}
}
}
}
\usage{
reference_chromosomes
}
\description{
Correspondences between the assembly and alignment names for the reference
species.
}
\keyword{datasets}
