% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rename_chrom_for_one_species}
\alias{rename_chrom_for_one_species}
\title{Rename the chromosomes for one species}
\usage{
rename_chrom_for_one_species(
  species_name,
  temp_names_chroms,
  chromosome_correspondences,
  all_alignments,
  ref_species
)
}
\arguments{
\item{species_name}{This string contains the names of the focal species.}

\item{temp_names_chroms}{This data frame contains the names of the chromosomes
for the focal species compared to the reference.}

\item{chromosome_correspondences}{This data frame contains the
correspondences between the chromosomes of the focal species and the
reference species.}

\item{all_alignments}{This data frame contains all the coordinates of the aligned
genomes on one another.}

\item{ref_species}{Name of the species used as reference for the analysis.}
}
\value{
A data frame containing the new names of the chromosomes for the
focal species.
}
\description{
This function filters the input data for the specified species and
renames the chromosomes for this species. First it renames chromosomes
that are fissionned (the chromosome is split in two in the reference
species and is only one chromosome in the focal species). Then, it renames
chromosomes that are fused (chromosomes that are split in two in the
reference species and fused in the focal species).
}
\examples{
# Import the test data
data(example_rename_chromosomes)
example_rename_chromosomes
data(correspondences_chromosomes)
correspondences_chromosomes
data(all_alignments)
all_alignments

species_name <- unique(example_rename_chromosomes$Species)[2]

rename_chrom_for_one_species(species_name, example_rename_chromosomes,
                             correspondences_chromosomes, all_alignments)
}
