% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rename_fissionned_chroms}
\alias{rename_fissionned_chroms}
\title{Rename chromosomes that are fissionned (one chromosome in the reference
species and split in two chromosomes in the focal species).}
\usage{
rename_fissionned_chroms(
  chromosome_name,
  temp_names_for_sp,
  all_alignments,
  chromosome_correspondences,
  ref_species
)
}
\arguments{
\item{chromosome_name}{This string is the name that was given to the chromosome
in the analysis}

\item{temp_names_for_sp}{This data frame contains all the names of the
chromosomes for the focal species}

\item{all_alignments}{This data frame contains all the coordinates of the
mapping sequences throughout the genome. It is the output of the \code{input_paf}
function.}

\item{chromosome_correspondences}{This data frame contains the
correspondences between the chromosomes of the focal species and the
reference species.}

\item{ref_species}{Name of the species used as reference to name the chromosomes.}
}
\value{
A data frame containing the new names of the fissionned chromosomes
}
\description{
This function keeps all the lines in the data frame containing the name of
the chromosome and uses the \code{rename_portions_of_chromosomes_fusion} function
to rename these portions of chromosomes.
}
\examples{
# Import the test data
data(example_rename_chromosomes)
example_rename_chromosomes
data(correspondences_chromosomes)
correspondences_chromosomes
data(all_alignments)
all_alignments

species_name <- unique(example_rename_chromosomes$Species)[2]

temp_names_for_sp <- example_rename_chromosomes[
 which(example_rename_chromosomes$Species == species_name),
]

rename_fissionned_chroms("LG_3", temp_names_for_sp, all_alignments,
                        correspondences_chromosomes)

}
