% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rename_fused_chromosomes}
\alias{rename_fused_chromosomes}
\title{Rename fused chromosomes (two chromosomes in the reference species and one
chromosome in the focal species).}
\usage{
rename_fused_chromosomes(assembly_name, temp_names_for_sp)
}
\arguments{
\item{assembly_name}{This string is the name of the chromosome given during
the assembly.}

\item{temp_names_for_sp}{This data frame contains the names of all the
chromosomes for the focal species}
}
\value{
A data frame containing the new names of the chromosomes.
}
\description{
This function filters the table of all the names given to the chromosomes
using their assembly names. If there are more than one name given to the
same assembly chromosome, we consider it to be a fusion (one chromosome
in the reference species and tow in the focal species) and rename it to have
the same name.
}
\examples{
# Import the test data
data(example_rename_chromosomes)
example_rename_chromosomes

species_name <- unique(example_rename_chromosomes$Species)[2]

temp_names_for_sp <- example_rename_chromosomes[
 which(example_rename_chromosomes$Species == species_name),
]

temp_names_for_sp <- dplyr::mutate(temp_names_for_sp,
 Chromosome_name = dplyr::case_when(
   Chromosome == "OZ349660" & Chromosome_name == "LG_9"  ~ "LG_9.a",
   Chromosome == "OZ349660" & Chromosome_name == "LG_10"  ~ "LG_9.b",
   TRUE ~ Chromosome_name
   )
)

rename_fused_chromosomes("OZ349660", temp_names_for_sp)
}
