% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rename_portions_of_chromosomes_fusion}
\alias{rename_portions_of_chromosomes_fusion}
\title{Reorder and rename the portions of chromosomes that are in different orders
in different species (especially the case for fusions/fissions of chromosomes).}
\usage{
rename_portions_of_chromosomes_fusion(
  temp_names_for_sp_for_chr,
  all_alignments,
  chromosome_correspondences,
  ref_species
)
}
\arguments{
\item{temp_names_for_sp_for_chr}{This data frame contains the names of the
chromosomes of species A and B, their respective mapping positions and the
names of species A and B.}

\item{all_alignments}{This data frame contains all the alignments between the
two species.}

\item{chromosome_correspondences}{This data frame contains the correspondences
between chromosomes of the two species. It is the output of the \code{get_corresponding_chromosomes_species}
function.}

\item{ref_species}{Name of the species used as reference to name the chromosomes.}
}
\value{
A data frame containing the new names of the chromosomes
}
\description{
This function computes the mean mapping position of all the chromosomes of
species B that map on one chromosome of species A in the case of fusions.
It reorders the fragments of the fractionned chromosome for them to be in
accordance with the chromosome of species A. For example, if we have LG1.a-b
in species A that is separated into LG1.a and LG1.b in species B and that
the portion called LG1.a maps with a portion of LG1.a-b that is located
after the portion of LG1.a-b that maps with LG1.b, this function will
swap the names of LG1.a and LG1.b so that the names make sense with the
mapping.
}
\examples{
# Import the test data
data(example_rename_chromosomes)
example_rename_chromosomes
data(correspondences_chromosomes)
correspondences_chromosomes
data(all_alignments)
all_alignments

species_name <- unique(example_rename_chromosomes$Species)[2]
chrom_name <- "LG_3"

condition_1 <- example_rename_chromosomes$Species == species_name
condition_2 <- example_rename_chromosomes$Chromosome_name == chrom_name

temp_names_for_sp_for_chr <- example_rename_chromosomes[
 which(condition_1 & condition_2),
]

rename_portions_of_chromosomes_fusion(temp_names_for_sp_for_chr,
                                      all_alignments,
                                      correspondences_chromosomes)
}
