% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{return_plot}
\alias{return_plot}
\title{Returns a plot if there is no specified file name}
\usage{
return_plot(file_name, p)
}
\arguments{
\item{file_name}{This is the name of the file in which to save the plot}

\item{p}{This is a ggplot object}
}
\value{
Either a plot or a message depending on the specified file name.
}
\description{
This function checks if there is a specified file name. If it is the case
it saves the plot at the specified location. If not, it returns the plot.
}
\examples{
# Prepare data
df <- data.frame(
  "x" = seq(1, 10),
  "y" = seq(1, 20, 2)
)

p <- ggplot2::ggplot(df,
                     ggplot2::aes(x = x, y = y)) +
     ggplot2::geom_line(lwd = 1.2)
return_plot(NULL, p)

# Save the plot in a file
# return_plot("test_return_plot.png", p)
}
