% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{reverse_all_chromosomes}
\alias{reverse_all_chromosomes}
\title{Reverse chromosomes that are not oriented in the same way in all the reference
genomes}
\usage{
reverse_all_chromosomes(all_alignments, correspondences_chromosomes)
}
\arguments{
\item{all_alignments}{This data frame is the output of the paf alignments
transformed into a table.}

\item{correspondences_chromosomes}{This data frame contains the correspondences
of the chromosomes from one species to the other. It is the output of the
\code{get_corresponding_chromosomes_species} function.}
}
\value{
A similar data frame as the \code{all_alignments} input, but with
chromosomes aligned in the same direction in the different species.
}
\description{
This function goes through all the pairs of species and the pairs of
chromosomes, checks if they are reversed compared to each other
and reverses one of the two chromosomes if it is not already done.
}
\examples{
data(all_alignments)

data(names_of_chromosomes)

correspondences_chromosomes <- get_corresponding_chromosomes_species(
   all_alignments, names_of_chromosomes
)

reverse_all_chromosomes(all_alignments, correspondences_chromosomes)
}
