% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{reverse_chromosome}
\alias{reverse_chromosome}
\title{Reverse the order of the positions along a specified chromosome.}
\usage{
reverse_chromosome(all_alignments, name_chromosome_to_turn)
}
\arguments{
\item{all_alignments}{This data frame contains all the coordinates of the aligned
genomes on one another.}

\item{name_chromosome_to_turn}{This string is the name that was given to the
chromosome to reorder when it was assembled.}
}
\value{
A data frame similar to the input \code{all_data}, but with the
positions of the focal chromosome reversed.
}
\description{
This function determines the maximum values of a chromosome
given its name and rearranges the positions of this chromosome by defining
its last position as its first and its first position as its last.
}
\examples{
data(all_alignments)

reverse_chromosome(all_alignments, "OZ239481")
}
