% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{trace_dotplot}
\alias{trace_dotplot}
\title{Trace a dotplot.}
\usage{
trace_dotplot(
  alignments,
  name_of_chromosomes,
  only_chromosome = NULL,
  small_alignment_threshold = 50000,
  file_name = NULL
)
}
\arguments{
\item{alignments}{This table contains the alignments for the different species
combinations. It is the equivalent of the \code{.paf} files, but transformed as tables}

\item{name_of_chromosomes}{This table contains the names of the chromosomes
for the different species. It is the output of the \code{get_chromosome_names_all_species}
function}

\item{only_chromosome}{This argument is a string or vector of strings that
contains the name of one chromosome if you want to look at it individually.
It allows to filter the data to look only at this chromosome. The default
value is \code{NULL}.}

\item{small_alignment_threshold}{This number allows to filter small fragments
that map. It defines the threshold of what the user specifies as "small". The
default value is \code{5e4}}

\item{file_name}{This string contains the name of the file to be produced in
your working directory. The default is \code{NULL}. In this case, it plots the graph
in your plot viewer. As soon as a string is specified, it will save the plot
without plotting it.}
}
\value{
The dotplot as a ggplot2 object
}
\description{
This function allows to trace a dotplot between the alignments of two
species if you want to look at just one pair of species.
}
\examples{
# Import data
data(alignments_ready_to_plot)
data(chromosome_names)

# Test the function
trace_dotplot(alignments_ready_to_plot,
              chromosome_names)

# You can also choose to save the plot directly
# trace_dotplot(alignments_ready_to_plot,
#               chromosome_names,
#               file_name = "test.png")

# You can also choose to plot only one chromosome
# trace_dotplot(alignments_ready_to_plot,
#               chromosome_names,
#               only_chromosome = "LG_1$",
#               file_name = "test_one_chromosome.png")
}
