% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{which_not_in}
\alias{which_not_in}
\title{Function to get the values that are in x and not in y}
\usage{
which_not_in(x, y)
}
\arguments{
\item{x}{This vector is the vector to check}

\item{y}{This vector is the one being checked}
}
\value{
A vector containing the values that are in x and not in y
}
\description{
This function uses the \code{which} and the \verb{\%not_in\%} functions to keep the
values that are in x and not in y.
}
\examples{
vect1 <- c("a", "b", "x", "c")
vect2 <- c("a", "b", "c", "d")
which_not_in(vect1, vect2)
}
