% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{chrom_name_contains_pattern}
\alias{chrom_name_contains_pattern}
\title{Checks if the chromosome name contains a specified pattern}
\usage{
chrom_name_contains_pattern(df, pattern)
}
\arguments{
\item{df}{This data frame has to contain the column \code{Chromosome}}

\item{pattern}{This string is the pattern to look for in the \code{Chromosome}
column}
}
\value{
Boolean: \code{TRUE} if the chromosome name contains the pattern and
\code{FALSE} otherwise
}
\description{
This function uses the \code{grepl} function to check if the specified pattern
is inside the name of the chromosomes (taken from the \code{Chromosome} column
from the specified data frame)
}
\examples{
df1 <- data.frame("Chromosome" = c("Chrom_1", "Chrom_2", "Chrom_3"), "Length" = c(10, 20, 250))
df1
chrom_name_contains_pattern(df1, "_")
}
