% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{define_reference_for_chromosome_naming}
\alias{define_reference_for_chromosome_naming}
\title{Define a reference species and chromosome names for the chromosome naming}
\usage{
define_reference_for_chromosome_naming(all_chromosomes, prefix = "Chrom")
}
\arguments{
\item{all_chromosomes}{This data frame contains at least three columns:
the name of the chromosomes that was given to them when they were
assembled, the length of the chromosomes and the species to which the
chromosomes belong.}

\item{prefix}{This string is the beginning of the name of the chromosomes.
Its default is "Chrom" to have "Chrom_1", ..., but it can be changed to
any string. Usual ones are "Chr", "Chrom" or "LG".}
}
\value{
A table containing three columns: The name of the chromosome that
was given when it was assembled for the species chosen as reference, the new
name of the chromosome and the species that was chosen as reference.
}
\description{
This function takes the first species alphabetically in the
table of all chromosomes from the different species, orders the chromosomes
from the longest to the shortest and names them from "Chrom_1" to "Chrom_n",
n being the number of chromosomes that are given for the species chosen as
reference.
}
\examples{
# Import the assembly names and lengths of chromosomes
data(names_of_chromosomes)

define_reference_for_chromosome_naming(names_of_chromosomes)
define_reference_for_chromosome_naming(names_of_chromosomes, prefix = "LG")
}
