% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{import_chromosome_length}
\alias{import_chromosome_length}
\title{Import the names and lengths of chromosome for one species}
\usage{
import_chromosome_length(path)
}
\arguments{
\item{path}{Path to where the file is stored.}
}
\value{
A data frame containing the name of the chromosome in one column
and its length in the other.
}
\description{
This function allows to import the names and lengths of chromosomes from
".dict" or ".fai" files. It also allows to simplify names of chromosomes
that might have a dot or a vertical line ("|").
}
\examples{
fai_file <- "TidySynt/extdata/refgenome/Jaera_albifrons_chromosomes.fasta.fai"
dict_file <- "TidySynt/extdata/refgenome/Jaera_albifrons_chromosomes.dict"

import_chromosome_length(fai_file)
import_chromosome_length(dict_file)
}
