% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{import_paf_from_folder}
\alias{import_paf_from_folder}
\title{Import all paf files from a given directory and transform them into a tidy
table.}
\usage{
import_paf_from_folder(dir_name)
}
\arguments{
\item{dir_name}{A string containing the name of the directory in which to
look for \code{paf} files.}
}
\value{
A data frame containing all the stats and coordinates of the mapping
sequences along the chromosomes of the query and target species.
}
\description{
This function makes a list of all the \code{paf} files in the provided directory,
gets the names of the species (query and target) and imports them all. A
supplementary step is that the names of the species (query and target) are
added to the table.
}
\examples{
import_paf_from_folder("TidySynt/extdata/paf/")
}
