% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rename_chromosomes_after_calling}
\alias{rename_chromosomes_after_calling}
\title{This function allows to rename the chromosomes after they were named. This
is especially useful for fusion/fissions and translocations.}
\usage{
rename_chromosomes_after_calling(
  temp_names_chroms,
  chromosome_correspondences,
  all_alignments
)
}
\arguments{
\item{temp_names_chroms}{This data frame contains the names of the chromosomes
that were attributed using the maximum cumulative mapping sequence.}

\item{chromosome_correspondences}{This data frame contains the
correspondences between the chromosomes of the focal species and the
reference species.}

\item{all_alignments}{This data frame contains all the coordinates of the aligned
genomes on one another.}
}
\value{
A data frame that contains the new names of the chromosomes for all
chromosomes for all species.
}
\description{
This function uses as input the names that were given to the chromosomes
using only corresponding chromosomes. For each chromosome, it checks if
it is a rearrangement and if it is, it renames the portions of the
rearrangements. If not, it simply continues.
}
\examples{
# Import the test data
data(example_rename_chromosomes)
example_rename_chromosomes
data(correspondences_chromosomes)
correspondences_chromosomes
data(all_alignments)
all_alignments

rename_chromosomes_after_calling(example_rename_chromosomes,
                                 correspondences_chromosomes,
                                 all_alignments)
}
